requirements.txt
torch
torchvision
mlflow
numpy
matplotlib
Pillow
scikit-learn
